Separate between the spring application and its controller classes (#49)
* move practitioner controller into its own package

* update sample-service structure

* small update on the PR template
