+++ .github/pull_request_template.md
@@ -2,4 +2,9 @@
 Please replace this text with a summary of the change and which issue is fixed.
 
 ### Changes
-- ARTS-### - some description
+- [ARTS-###] - some description
+
+### Checklist:
+
+- [ ] Branch name follows convention (feature/arts-###-short-name OR fix/arts-###-short-name)
+- [ ] I have included a short-meaningful title, description and changes for this PR
