+++ sample-service/src/main/java/uk/ac/ox/ndph/mts/sample_service/controller/MainController.java
@@ -1,9 +1,8 @@
-package uk.ac.ox.ndph.mts.sample_service;
+package uk.ac.ox.ndph.mts.sample_service.controller;
 
 import uk.ac.ox.ndph.mts.sample_service.config.ConfigService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -11,15 +10,15 @@ import org.springframework.web.bind.annotation.RestController;
  * The application
  */
 @RestController
-@SpringBootApplication
-public class DemoApp {
+public class MainController {
 
     /**
      * The constructor for this class.
+     *
      * @param configService a config service
      */
     @Autowired
-    public DemoApp(ConfigService configService) {
+    public MainController(ConfigService configService) {
         this.myConfig = configService;
     }
 
@@ -30,6 +29,7 @@ public class DemoApp {
 
     /**
      * The endpoint to get the secret.
+     *
      * @return the secret as a string.
      */
     @GetMapping("/getsecret")
@@ -39,6 +39,7 @@ public class DemoApp {
 
     /**
      * An additional endpoint.
+     *
      * @return a message as a string.
      */
     @GetMapping("/hello")
