+++ sample-service/src/main/java/uk/ac/ox/ndph/mts/sample_service/controller/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * The controllers package of this application
+ */
+package uk.ac.ox.ndph.mts.sample_service.controller;
